backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
ultralytics>=8.0.0
pillow>=9.0.0
numpy>=1.24.0
python-multipart>=0.0.6
